+++ FluentConsole.cs
@@ -23,7 +23,6 @@ namespace FluentConsoleNet
 
         public static FluentConsoleBuilder Console => FluentConsoleBuilder.Create(Logger);
 
-        public static readonly FastConsole FastConsole = new FastConsole();
 
         public static void AttachLogger(dynamic logger)
         {
@@ -33,6 +32,14 @@ namespace FluentConsoleNet
             }
 
             Logger = logger;
+
+        }
+
+
+        public static void DetachLogger()
+        {
+
+            Logger = null;
 
         }